feat: mui right to left configuration and install packages prop-types
