+++ client/.eslintrc.json
@@ -18,6 +18,9 @@
         "react"
     ],
     "rules": {
+        // suppress errors for missing 'import React' in files
+        "react/react-in-jsx-scope": "off",
+        // allow jsx syntax in js files (for next.js project)
         "react/jsx-filename-extension": [1, { "extensions": [".js", ".jsx"] }]
     },
     "root": true
